fixed some deprecated stuff

+++ evaluate.py
@@ -9,10 +9,13 @@ import torch
 import torch.nn as nn
 from torch.utils import data
 from generate_keypoints import process_video
-from models import Transformer
-from configs import TransformerConfig
+from models import LSTM
+from configs import LstmConfig
 from utils import load_json, load_label_map
 import shutil
+
+# RUN THIS:
+# python evaluate.py --data_dir D:/dev/datasets/eval_include/
 
 parser = argparse.ArgumentParser(description="Evaluate function")
 parser.add_argument("--data_dir", required=True, help="data directory")
@@ -56,7 +59,11 @@ class KeypointsDataset(data.Dataset):
         x, y = np.array(x), np.array(y)
         _, length = x.shape
         x = x.reshape((-1, length, 1))
+        # print(x.shape)
         y = y.reshape((-1, length, 1))
+        # print(y.shape)
+        # res = np.concatenate((x, y), -1).astype(np.float32)
+        # print(res.shape)
         return np.concatenate((x, y), -1).astype(np.float32)
 
     def __getitem__(self, idx):
@@ -122,48 +129,52 @@ def inference(dataloader, model, device, label_map):
         input_data = batch["data"].to(device)
         output = model(input_data).detach().cpu()
         output = torch.argmax(torch.softmax(output, dim=-1), dim=-1).numpy()
-        predictions.append({"uid": batch["uid"][0], "predicted_label": label_map[output[0]]})
+        predictions.append(
+            {"uid": batch["uid"][0], "predicted_label": label_map[output[0]]}
+        )
 
     return predictions
 
 
-video_paths = glob.glob(os.path.join(args.data_dir, "*"))
-save_dir = "keypoints_dir"
-if os.path.isdir(save_dir):
-    shutil.rmtree(save_dir)
-os.mkdir(save_dir)
-for path in tqdm(video_paths, desc="Processing Videos"):
-    process_video(path, save_dir)
+if __name__ == "__main__":
+    video_paths = glob.glob(os.path.join(args.data_dir, "*"))
+    save_dir = "keypoints_dir"
+    already_processed_videos = True
+    if not already_processed_videos:
+        if os.path.isdir(save_dir):
+            shutil.rmtree(save_dir)
+        os.mkdir(save_dir)
+        for path in tqdm(video_paths, desc="Processing Videos"):
+            process_video(path, save_dir)
 
-label_map = load_label_map("include")
-dataset = KeypointsDataset(
-    keypoints_dir=save_dir,
-    max_frame_len=169,
-)
+    label_map = load_label_map("include")
+    dataset = KeypointsDataset(
+        keypoints_dir=save_dir,
+        max_frame_len=169,
+    )
 
-dataloader = data.DataLoader(
-    dataset,
-    batch_size=1,
-    shuffle=False,
-    num_workers=4,
-    pin_memory=True,
-)
-label_map = dict(zip(label_map.values(), label_map.keys()))
+    dataloader = data.DataLoader(
+        dataset,
+        batch_size=1,
+        shuffle=False,
+        num_workers=4,
+        pin_memory=True,
+    )
+    label_map = dict(zip(label_map.values(), label_map.keys()))
 
-device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-config = TransformerConfig(size="large", max_position_embeddings=256)
-model = Transformer(config=config, n_classes=263)
-model = model.to(device)
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    config = LstmConfig()
+    model = LSTM(config=config, n_classes=263)
+    model = model.to(device)
 
-pretrained_model_name = "include_no_cnn_transformer_large.pth"
-pretrained_model_links = load_json("pretrained_links.json")
-if not os.path.isfile(pretrained_model_name):
-    link = pretrained_model_links[pretrained_model_name]
-    torch.hub.download_url_to_file(link, pretrained_model_name, progress=True)
+    pretrained_model_name = "include_no_cnn_lstm.pth"
+    pretrained_model_links = load_json("pretrained_links.json")
+    if not os.path.isfile(pretrained_model_name):
+        link = pretrained_model_links[pretrained_model_name]
+        torch.hub.download_url_to_file(link, pretrained_model_name, progress=True)
 
-ckpt = torch.load(pretrained_model_name)
-model.load_state_dict(ckpt["model"])
-print("### Model loaded ###")
-
-preds = inference(dataloader, model, device, label_map)
-print(json.dumps(preds, indent=2))
+    ckpt = torch.load(pretrained_model_name)
+    model.load_state_dict(ckpt["model"])
+    print("### Model loaded ###")
+    preds = inference(dataloader, model, device, label_map)
+    print(json.dumps(preds, indent=2))
